refactor: elevate mounting points for play time editor and score editor dialogs
- Manage dialog states using zustand to avoid complex prop drilling
- Prepare for adding multiple dialog triggers in the future
