+++ src/renderer/src/components/Game/Config/ManageMenu/main.tsx
@@ -1,3 +1,7 @@
+import { useTranslation } from 'react-i18next'
+import { toast } from 'sonner'
+import { ipcManager } from '~/app/ipc'
+import { useLibrarybarStore } from '~/components/Librarybar/store'
 import {
   DropdownMenuGroup,
   DropdownMenuItem,
@@ -7,28 +11,17 @@ import {
   DropdownMenuSubContent,
   DropdownMenuSubTrigger
 } from '~/components/ui/dropdown-menu'
-import { useTranslation } from 'react-i18next'
-import { toast } from 'sonner'
-import { useGameLocalState, useGameState, useConfigState } from '~/hooks'
+import { useConfigState, useGameLocalState, useGameState } from '~/hooks'
 import { useGameAdderStore } from '~/pages/GameAdder/store'
 import { useGameDetailStore } from '../../store'
-import { useLibrarybarStore } from '~/components/Librarybar/store'
 import { DeleteGameAlert } from './DeleteGameAlert'
-import { ipcManager } from '~/app/ipc'
-import { useState } from 'react'
-import { Dialog, DialogContent } from '~/components/ui/dialog'
-import { Input } from '~/components/ui/input'
-import { Button } from '~/components/ui/button'
-import { cn } from '~/utils'
 
 export function ManageMenu({
   gameId,
-  openNameEditorDialog,
-  openPlayingTimeEditorDialog
+  openNameEditorDialog
 }: {
   gameId: string
   openNameEditorDialog: () => void
-  openPlayingTimeEditorDialog: () => void
 }): React.JSX.Element {
   const { t } = useTranslation('game')
   const [gamePath] = useGameLocalState(gameId, 'path.gamePath')
@@ -36,64 +29,24 @@ export function ManageMenu({
   const [gameName] = useGameState(gameId, 'metadata.name')
   const [nsfw, setNsfw] = useGameState(gameId, 'apperance.nsfw')
   const [playStatus, setPlayStatus] = useGameState(gameId, 'record.playStatus')
-  const [score, setScore] = useGameState(gameId, 'record.score')
-  const [preScore, setPreScore] = useState(score === -1 ? '' : score.toString())
   const [selectedGroup] = useConfigState('game.gameList.selectedGroup')
-  const [isScoreDialogOpen, setIsScoreDialogOpen] = useState(false)
   const { refreshGameList } = useLibrarybarStore.getState()
   const setIsEditingLogo = useGameDetailStore((state) => state.setIsEditingLogo)
+  const setIsPlayTimeEditorDialogOpen = useGameDetailStore(
+    (state) => state.setIsPlayTimeEditorDialogOpen
+  )
+  const setIsRatingEditorDialogOpen = useGameDetailStore(
+    (state) => state.setIsRatingEditorDialogOpen
+  )
   const setIsOpen = useGameAdderStore((state) => state.setIsOpen)
   const setName = useGameAdderStore((state) => state.setName)
   const setDbId = useGameAdderStore((state) => state.setDbId)
-
-  const resetPreScore = (): void => setPreScore(score === -1 ? '' : score.toString())
 
   const changePlayStatus = (value: typeof playStatus): void => {
     setPlayStatus(value)
     if (selectedGroup === 'record.playStatus') {
       refreshGameList()
     }
-  }
-
-  // Score submission handler function
-  function confirmScore(): void {
-    if (preScore === '') {
-      setScore(-1)
-      setIsScoreDialogOpen(false)
-      toast.success(t('detail.header.rating.cleared'))
-      return
-    }
-
-    const scoreNum = parseFloat(preScore)
-
-    if (isNaN(scoreNum)) {
-      toast.error(t('detail.header.rating.errors.invalidNumber'))
-      resetPreScore()
-      return
-    }
-
-    if (scoreNum < 0) {
-      toast.error(t('detail.header.rating.errors.negative'))
-      resetPreScore()
-      return
-    }
-
-    if (scoreNum > 10) {
-      toast.error(t('detail.header.rating.errors.tooHigh'))
-      resetPreScore()
-      return
-    }
-
-    const formattedScore = scoreNum.toFixed(1)
-
-    if (preScore !== formattedScore && !Number.isInteger(scoreNum)) {
-      toast.warning(t('detail.header.rating.warning'))
-    }
-
-    setScore(Number(formattedScore))
-    setPreScore(Number(formattedScore).toString())
-    setIsScoreDialogOpen(false)
-    toast.success(t('detail.header.rating.success'))
   }
 
   return (
@@ -114,7 +67,7 @@ export function ManageMenu({
                 {t('detail.manage.editLogo')}
               </DropdownMenuItem>
 
-              <DropdownMenuItem onClick={openPlayingTimeEditorDialog}>
+              <DropdownMenuItem onClick={() => setIsPlayTimeEditorDialogOpen(true)}>
                 {t('detail.manage.editPlayTime')}
               </DropdownMenuItem>
 
@@ -162,13 +115,7 @@ export function ManageMenu({
               </DropdownMenuSub>
 
               {/* Rating menu item */}
-              <DropdownMenuItem
-                onSelect={(e) => {
-                  e.preventDefault()
-                  resetPreScore()
-                  setIsScoreDialogOpen(true)
-                }}
-              >
+              <DropdownMenuItem onSelect={() => setIsRatingEditorDialogOpen(true)}>
                 {t('detail.header.rating.tooltip')}
               </DropdownMenuItem>
 
@@ -228,23 +175,6 @@ export function ManageMenu({
           </DropdownMenuPortal>
         </DropdownMenuSub>
       </DropdownMenuGroup>
-
-      {/* Rating dialog */}
-      <Dialog open={isScoreDialogOpen} onOpenChange={setIsScoreDialogOpen}>
-        <DialogContent showCloseButton={false} className="w-[500px]">
-          <div className={cn('flex flex-row gap-3 items-center justify-center')}>
-            <div className={cn('whitespace-nowrap')}>{t('detail.header.rating.title')}</div>
-            <Input
-              value={preScore}
-              onChange={(e) => setPreScore(e.target.value)}
-              onKeyDown={(e) => {
-                if (e.key === 'Enter') confirmScore()
-              }}
-            />
-            <Button onClick={confirmScore}>{t('utils:common.confirm')}</Button>
-          </div>
-        </DialogContent>
-      </Dialog>
     </>
   )
 }
